requirements.txt
asgiref==3.8.1
Django==5.0.4
django-model-utils==4.5.0
djangorestframework==3.15.1
greenlet==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
